+++ create.sh
@@ -8,6 +8,10 @@ cd ..
 cd windows
 upx display_wenservice_windows.exe
 cd ..
-docker rmi -f petrjahoda/display_webservice:"$1"
-docker build -t petrjahoda/display_webservice:"$1" .
-docker push petrjahoda/display_webservice:"$1"
+docker rmi -f petrjahoda/display_webservice:latest
+docker build -t petrjahoda/display_webservice:latest .
+docker push petrjahoda/display_webservice:latest
+
+docker rmi -f petrjahoda/display_webservice:2020.1.3
+docker build -t petrjahoda/display_webservice:2020.1.3 .
+docker push petrjahoda/display_webservice:2020.1.3