+++ CHANGELOG.md
@@ -14,6 +14,13 @@ adhere to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) rules.
 * ```Fixed``` for any bug fixes.
 * ```Security``` in case of vulnerabilities.
 
+## [2020.1.3.31] - 2020-03-31
+
+### Added
+- updated create.sh for uploading proper docker version automatically
+
+### Fixed
+- proper sending time to display
 
 ## [2020.1.3.5] - 2020-03-05